+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * refactor(swarm): swarm cleanup following libp2p upgrade to v0.39.1 [#473]
 * fix: strict ordering for DAG-CBOR-encoded map keys [#493]
 * feat: upgrade libp2p to v0.43.0 [#499]
+* feat(http): default values for --bits and --profile [#502]
 
 [#429]: https://github.com/rs-ipfs/rust-ipfs/pull/429
 [#428]: https://github.com/rs-ipfs/rust-ipfs/pull/428
@@ -31,6 +32,7 @@
 [#473]: https://github.com/rs-ipfs/rust-ipfs/pull/473
 [#493]: https://github.com/rs-ipfs/rust-ipfs/pull/493
 [#499]: https://github.com/rs-ipfs/rust-ipfs/pull/499
+[#502]: https://github.com/rs-ipfs/rust-ipfs/pull/502
 
 # 0.2.1